[Test] Refactor quantization UT module (#16084)
### What this PR does / why we need it?

**Purpose**

- Align quantization UT directory with the refactored
`vllm_ascend/quantization/` module structure.
- Separate mixed CPU (mock-based) and NPU test cases under
`tests/e2e/pull_request/`.

**Approach**

1. **Directory alignment**: Reorganized `tests/ut/quantization/` into
`configs/` and `methods/{kv_cache, w4a4, w4a8, w8a8, wna16}`, mirroring
the source layout.
2. **Test separation**: Moved CPU mock-based cases into the
corresponding `tests/ut/quantization/methods/<method>/`; consolidated
NPU-dependent cases into
`tests/e2e/pull_request/one_card/test_linear_quantization.py`.

### Does this PR introduce _any_ user-facing change?

no

### How was this patch tested?

UT (CPU only): `pytest -sv tests/ut/quantization/`
E2E (NPU required): `pytest -sv
tests/e2e/pull_request/one_card/test_linear_quantization.py`


- vLLM main:
https://github.com/vllm-project/vllm/commit/b2f685834a6456197e7033966fdef52a23f1abcd
