+++ .github/workflows/scripts/test_config.yaml
@@ -223,11 +223,7 @@ estimated_times:
   tests/e2e/pull_request/one_card/test_gdn_layerwise_kv.py: 70
   tests/e2e/pull_request/two_card/test_activation.py: 50
   tests/e2e/pull_request/two_card/test_kimi_kda_fused_norm_gate.py: 60
-  tests/e2e/pull_request/one_card/test_w4a4_flatquant.py: 40
-  tests/e2e/pull_request/one_card/test_w4a4_laos_dynamic.py: 40
-  tests/e2e/pull_request/one_card/test_w8a16.py: 40
-  tests/e2e/pull_request/one_card/test_w8a8_dynamic.py: 30
-  tests/e2e/pull_request/one_card/test_w8a8_static.py: 40
+  tests/e2e/pull_request/one_card/test_linear_quantization.py: 60
   tests/e2e/pull_request/one_card/test_gumbel_sampling.py: 110
   tests/e2e/pull_request/one_card/test_model_runner_v1_with_device.py: 70
