+++ tests/ut/quantization/methods/w8a8/test_w8a8_static.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock, Mock, patch
 import torch
 
 from tests.ut.base import TestBase
-from tests.ut.quantization.conftest_quantization import create_linear_layer, identity
+from tests.ut.quantization.conftest_quantization import identity
 from vllm_ascend.quantization.methods.w8a8.w8a8_static import AscendW8A8LinearMethod
 from vllm_ascend.utils import ASCEND_QUANTIZATION_METHOD, COMPRESSED_TENSORS_METHOD
 
@@ -155,34 +155,3 @@ class TestAscendW8A8LinearMethod(TestBase):
         self.assertEqual(layer.weight_offset.data.shape, (128,))
         mock_npu_format_cast.assert_called_once()
         self.assertFalse(isinstance(layer.deq_scale, MagicMock))
-
-
-class TestAscendW8A8LinearMethodWithNpu(TestBase):
-    @patch("vllm_ascend.quantization.methods.w8a8.w8a8_static.get_current_vllm_config")
-    def setUp(self, get_current_vllm_config):
-        mock_vllm_config = Mock()
-        mock_vllm_config.quant_config = Mock()
-        mock_vllm_config.quant_config.get_name.return_value = ASCEND_QUANTIZATION_METHOD
-        get_current_vllm_config.return_value = mock_vllm_config
-        self.method = AscendW8A8LinearMethod()
-        self.mock_get_config = patch("vllm_ascend.utils.get_ascend_config")
-        mock_config = self.mock_get_config.start()
-        mock_ascend_config = MagicMock()
-        mock_ascend_config.weight_nz_mode = 0
-        mock_config.return_value = mock_ascend_config
-
-    def tearDown(self):
-        self.mock_get_config.stop()
-
-    def test_apply_with_npu(self):
-        input_size, output_size = 128, 256
-        params_dtype = torch.bfloat16
-        layer = create_linear_layer(self.method, input_size, output_size, params_dtype)
-        layer.params_dtype = params_dtype
-        self.method.process_weights_after_loading(layer)
-
-        x = torch.randn(32, input_size, dtype=params_dtype).npu()
-        bias = torch.randn(output_size, dtype=torch.float32).npu()
-
-        output = self.method.apply(layer, x, bias)
-        self.assertEqual(output.shape, (32, output_size))
