+++ tests/ut/quantization/methods/wna16/test_w8a16.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock, patch
 import torch
 
 from tests.ut.base import TestBase
-from tests.ut.quantization.conftest_quantization import create_linear_layer, identity
+from tests.ut.quantization.conftest_quantization import identity
 from vllm_ascend.quantization.methods.wna16.w8a16 import AscendW8A16LinearMethod
 
 
@@ -67,28 +67,3 @@ class TestAscendW8A16LinearMethod(TestBase):
         self.assertEqual(layer.weight_scale.data.shape, (128,))
         self.assertEqual(layer.weight_offset.data.shape, (128,))
         mock_npu_format_cast.assert_called_once()
-
-
-class TestAscendW8A16LinearMethodWithNpu(TestBase):
-    def setUp(self):
-        self.method = AscendW8A16LinearMethod()
-        self.mock_get_config = patch("vllm_ascend.utils.get_ascend_config")
-        mock_config = self.mock_get_config.start()
-        mock_ascend_config = MagicMock()
-        mock_ascend_config.weight_nz_mode = 0
-        mock_config.return_value = mock_ascend_config
-
-    def tearDown(self):
-        self.mock_get_config.stop()
-
-    def test_apply_with_npu(self):
-        input_size, output_size = 128, 256
-        params_dtype = torch.bfloat16
-        layer = create_linear_layer(self.method, input_size, output_size, params_dtype)
-        self.method.process_weights_after_loading(layer)
-
-        x = torch.randn(32, input_size, dtype=params_dtype).npu()
-        bias = torch.randn(output_size, dtype=torch.float32).npu()
-
-        output = self.method.apply(layer, x, bias)
-        self.assertEqual(output.shape, (32, output_size))
